requirements.txt
# Requirements for many PaaS (as Heroku)
django
django-appconf==1.0.5
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-kn-defaults==0.3.1
django-oauth2-authcodeflow==0.4.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
pytest-django==4.5.2
mixer==7.2.0
pytest-cov==3.0.0
flake8==4.0.1
flake8-isort==4.1.1
black==21.12b0
pylint-django==2.5.0
pre-commit==2.16.0
factory-boy==3.2.1
djvue==0.2.1
pillow==8.3.1
django-cors-headers==3.7.0
sentry-sdk==1.5.1
elastic-apm==6.7.2
django-compressor==3.1
django-oauth2-authcodeflow==0.4.0
django-environ==0.8.1
psycopg2==2.9.5
gunicorn==20.0.4
bcrypt==4.0.1
rest_framework_simplejwt